Adding all the optimizations
