+++ Makefile
@@ -3,8 +3,7 @@
 all: check_format lint test
 
 lint:
-	pylint src/sr100_model_compiler/sr100_model_compiler.py
-	pylint tests
+	pylint src/sr100_model_compiler/sr100_model_compiler.py src/sr100_model_compiler/sr100_model_optimizer.py tests
 	#pylint src/sr100_model_compiler
 
 check_format:
