+++ src/sr100_model_compiler/__init__.py
@@ -3,6 +3,7 @@
 from .utils import shell_cmd
 from .utils import get_platform_path
 from .sr100_model_compiler import sr100_model_compiler
+from .sr100_model_optimizer import sr100_model_optimizer
 from .sr100_model_compiler import sr100_check_model
 from .sr100_model_compiler import sr100_get_compile_log
 
@@ -10,6 +11,7 @@ __all__ = [
     "shell_cmd",
     "get_platform_path",
     "sr100_model_compiler",
+    "sr100_model_optimizer",
     "sr100_check_model",
     "sr100_default_config",
 ]
